requirements.txt
django
djangorestframework
flake8
psycopg2
